requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
arabic-reshaper==2.1.3
async-timeout==5.0.1
attrs==26.1.0
beautifulsoup4==4.14.3
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
datasets==4.8.4
diffimg==0.2.3
dill==0.4.1
exceptiongroup==1.3.1
fastapi==0.136.1
filelock==3.29.0
frozenlist==1.8.0
fsspec==2026.2.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.47
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.13
Jinja2==3.1.6
jiwer==4.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.4.2
numpy==2.2.6
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.1
pandas==2.3.3
peft==0.19.1
pillow==12.2.0
platformdirs==4.9.6
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
pyarrow==24.0.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pyspellchecker==0.9.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
pytz==2026.1.post1
PyYAML==6.0.3
RapidFuzz==3.14.5
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
sentry-sdk==2.58.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
soupsieve==2.8.3
starlette==1.0.0
sympy==1.14.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.6.2
trdg==1.8.0
typer==0.24.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.46.0
wandb==0.26.1
wikipedia==1.4.0
xxhash==3.6.0
yarl==1.23.0
